Launch ForgotPasswordActivity in a new task.
This way the notificaiton about locked profile can function as
expected. Prior to this change ForgotPasswordActivity was launched
in the same task as initial profile credential confirmation. As a
result, when ForgotPasswordActivity was in the foreground, the
notification tap won't do anything since a task for
CONFIRM_DEVICE_CREDENTIAL_WITH_USER was already in the
foreground.

Bug: 157554361
Test: manual, with TestDPC
Change-Id: Ifc9d8e1fb8658c3f22926d7fa35da1e60eadb3db

Former-commit-id: dcd6fc593da212d985a37b199700211b29fccbc1