Improved UX when no debuggable apps are available.
In AppPicker, add an extra to exclude the "Nothing" sentinel. This is
used when selecting a debug app, and makes sense in that context.
However it does not makes sense when choosing an app to modify
compatibilty options for, so exclude it.

Also update the AppPicker to return a new result code in the case that
there are no apps available which match the criteria given. This result
code can only be used when the new extra is defined which keep this
change low risk.

In PlatformCompatDashboard, use the new extra and handle the new result
code, showing a dialog to explain that only debuggable apps are shown in
user builds. This also results in the "Nothing" item being shown at the
top of the app list.

Bug: 157633308
Test: Manual
Change-Id: Ifb055dd7c030cda42556bca8a9d7e87606f0ff72

Former-commit-id: 43f9adea620edad452210ce56712d16ce93bbb8b