Avoid double-badging application icon in the dialog
When creating a confirmation dialog for removing work profile, it will
fetch an icon from ApplicationInfo and badge the icon. However, the
cases of double-badging icon are happening in the some devices with
higher density, which means the icon has already had a badge after
getting it from ApplicationInfo. So this change attempts to directly use
the icon.

Bug: 148137991
Test: visual, manual
1. Install Test DPC from Play Store
2. Setup work profile
3. Settings -> Security -> Device admin apps
4. Click Test DPC item which has the badged icon -> Remove work profile
5. Observe the icon of the confirmation dialog
Change-Id: I126bb799e37e168fb5fbc76107c0c897a6be2f9a

Former-commit-id: e31fcb6dda510746baf2241c3dd78d2df9ef3a4f