Disable the defulat animation of the RecyclerView to avoid UI jank.
When users go to a sub page in Settings and come back to the homepage
again, those sticky cards flicker which makes the UI janky. This is
because we rebind sticky cards to ensure it's always up-to-date, where
the recyclerView has its default animation while notifyItemInserted(),
ane thus introduce the jank.

Fixes: 158627602
Test: Open Settings -> go to sub pages -> back to the homepage -> cards
are not flickering.

Change-Id: I2104dbe0bed8b2486c35521bcc0b5c8b54efb995

Former-commit-id: b54f9b4921547ea931ad1c939a6220279daa34c8