Add <uses-library> information.
The required <uses-library> "org.apache.http.legacy" is in the manifest
for Settings. It is necessary to explicitly add it to `uses_libs`
because it is not among the app dependencies (so Soong does not add it
automatically to the manifest_fixer). It also doesn't add it as a
compatibility library because the target API level for the app is not
below 28.

Bug: 132357300
Test: lunch cf_x86_phone-userdebug && m
Test: m out/soong/.intermediates/packages/apps/Settings/Settings/android_common/manifest_check/AndroidManifest.xml
Test: m out/soong/.intermediates/packages/apps/Settings/tests/robotests/SettingsRoboTestStub/android_common/manifest_check/AndroidManifest.xm
Change-Id: Id5d5fe1dff68cf6fce667a00e35710cf0e667d02

Former-commit-id: 202ca49dfed5e410d0f85f8c77570df0a6163c54