[Settings] Reduce the operation of VoLTE config redraw
There're 3 VoLTE configuration UIs sharing the same configuration,
but only 1 of them will be displayed.

The 3 classes extending Enhanced4gBasePreferenceController are:
1. Enhanced4gAdvancedCallingPreferenceController
2. Enhanced4gCallingPreferenceController
3. Enhanced4gLtePreferenceController
They're different in the value of getMode().

The design within Enhanced4gBasePreferenceController only allows to show one of them.
Which means, the other two don't need to register for PhoneStateListener and should ignore actions in response to it when receiving callback from PhoneStateListener.

Through detecting the VoLTE configuration mode, some operations could be
avoided.
This fix is targeting to avoid these operations from happening.

Bug: 172987488
Test: make RunSettingsRoboTests -j ROBOTEST_FILTER=Enhanced4gBasePreferenceControllerTest
Change-Id: I56d4900b5bcb189693c99e870b013da97ed283db

Former-commit-id: d2e22f4cee710ca34a964f5d03e36a92935120ed