Disable the title of mobile network
After the collapsing toolbar layout is applied to subsetting pages, the
mobile network page has two title shown currently. One is the activity
title while the other is the collapsing toolbar title To make sure
there will be only one title shown on this page, the activity title will
be removed and we will keep the collapsing toolbar title.

Bug: 176899590
Test: manual test and visual verified
1) Navigate to Settings -> Network & internet -> Mobile network
2) Check if there's only one title in this page

Change-Id: I50e57468c640f1d176fd99dcaafc06ed3fc601dc

Former-commit-id: 82df7e60a1fe82f6f9c8c8e38a36fac86ce1c3a7