Prevent a larger title from displaying incompletely
Some larger titles in the accessibility will be wrapped into 2 lines and
the tille looks like incomplete. This is because after starting to
scroll down, the collasping toolbar detected the length of title is
longer than the width of container and wrapped the title into 2 lines.

To prevent this, we intented to reduce the end margin for the collapsing
toolbar.

Fixes: 182235627
Test: visual verified
1) Adjust font size to largest
2) Navigate to Settings -> Accessibility -> Text and display -> Color
correction
3) Observe and see if the title is wrapped into 2 lines

Change-Id: I1595b9ab4f8406bf6505f56786438c9030b5be59

Former-commit-id: a5c91ef0fe8a9ad97230d66edeebd3730c1f3455