Fix no apps text overlapping with filter bar
Since Settings UI has been changed and the filter bar will be staying
in a fixed place, the "no apps" text will be overlapped with filter bar.
To make the text seperate from filter bar, the string will be put
in the bottom of text view.

Fixes: 182173848
Test: visual verified
1) Settings -> Notifications -> See all from last 7 days -> select
Turned off on filter bar
2) Observe and see if the "no apps" is overlapping with filter bar

Change-Id: Ib99bb5122fb72f05a2404b8f61b025a46d6d53dd

Former-commit-id: b0548113d27f2d13c0b52b286f47ec1974bdeb48