Don't always change username when adding account
During Setup Wizard for a new user, if an account is added,
the user's name gets updated to match their account's name.
Recently, however, the user can actually type in a name before
even reaching the Setup Wizard. If they elect to do so, we shouldn't
overwrite their name with their account name, so this cl makes
sure to only use the account name if the original name wasn't
manually entered.

Bug: 171255493
Test: manual verification (create new user and enter account during
      SUW. Do both with and without entering initial user name.)
Change-Id: I443648dafc0d83a71750f7f6472bbea4e2ee23bb

Former-commit-id: 0cb47e92d4bc1cf386e44dedadfe7d6b65e8f057